+++ nifi-commons/nifi-security-kerberos/src/main/java/org/apache/nifi/security/krb/TicketCacheConfiguration.java
