+++ nifi-commons/pom.xml
@@ -37,6 +37,8 @@
         <module>nifi-record-path</module>
         <module>nifi-rocksdb-utils</module>
         <module>nifi-schema-utils</module>
+        <module>nifi-security-kerberos</module>
+        <module>nifi-security-socket-ssl</module>
         <module>nifi-security-utils-api</module>
         <module>nifi-security-utils</module>
         <module>nifi-site-to-site-client</module>
