+++ nifi-nar-bundles/nifi-beats-bundle/nifi-beats-processors/pom.xml
@@ -36,6 +36,11 @@
             <version>1.14.0-SNAPSHOT</version>
         </dependency>
         <dependency>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-security-socket-ssl</artifactId>
+            <version>1.14.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
             <version>2.7</version>
