+++ nifi-nar-bundles/nifi-cassandra-bundle/nifi-cassandra-services/pom.xml
@@ -64,12 +64,6 @@
 
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-security-utils</artifactId>
-            <version>1.14.0-SNAPSHOT</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-mock</artifactId>
             <version>1.14.0-SNAPSHOT</version>
         </dependency>
