+++ nifi-nar-bundles/nifi-extension-utils/nifi-processor-utils/pom.xml
@@ -42,13 +42,18 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-security-utils</artifactId>
+            <artifactId>nifi-security-socket-ssl</artifactId>
             <version>1.14.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
             <version>2.6</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-lang3</artifactId>
+            <version>3.11</version>
         </dependency>
         <!-- Other modules using nifi-processor-utils are expected to have this API available, typically through a NAR dependency -->
         <dependency>
