+++ nifi-nar-bundles/nifi-kudu-bundle/nifi-kudu-processors/pom.xml
@@ -86,7 +86,12 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-security-utils</artifactId>
+            <artifactId>nifi-security-kerberos</artifactId>
+            <version>1.14.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-security-kerberos</artifactId>
             <version>1.14.0-SNAPSHOT</version>
         </dependency>
